Moved vision/los code to seperate class. continuing porting of seg032_C9D
