+++ src/player/playerdatcamera.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Data.Common;
 using System.Diagnostics;
 using Godot;
 
@@ -206,282 +207,5 @@ namespace Underworld
 
     }//end class
 
-    /// <summary>
-    /// Range of Vision Variables
-    /// </summary>
-    public class VisionParams : Loader
-    {
-        public static VisionParams[] visionparams = new VisionParams[0xF];
-        static byte[] _rawvisiondata = new byte[0xF * 0x11];
-        static short RelatedToFov_2C60 = 0;
-        public static short LikelyDistanceToWallOrDarkness = -1;
 
-        //array of data starting at dseg:2b5e
-        public byte dseg_2B5E
-        {
-            get
-            {
-                return (byte)getAt8(_rawvisiondata, ptr + 0);
-            }
-            set
-            {
-                setAt8(_rawvisiondata, ptr + 0, value);
-            }
-        }
-        public short FovYawX //= 0 //2B5F
-        {
-            get
-            {
-                return (short)getAt16(_rawvisiondata, ptr + 1);
-            }
-            set
-            {
-                setAt16(_rawvisiondata, ptr + 1, value);
-            }
-        }
-        public short FovYawY// = 0; //2B61
-        {
-            get
-            {
-                return (short)getAt16(_rawvisiondata, ptr + 3);
-            }
-            set
-            {
-                setAt16(_rawvisiondata, ptr + 3, value);
-            }
-        }
-
-        public short[] UNKARRAY_dseg_2B6B;
-
-        public byte dseg_2B63//;
-        {
-            get
-            {
-                return (byte)getAt8(_rawvisiondata, ptr + 5);
-            }
-            set
-            {
-                setAt8(_rawvisiondata, ptr + 5, value);
-            }
-        }
-        public byte CameraX_2b64//;//2B64
-        {
-            get
-            {
-                return (byte)getAt8(_rawvisiondata, ptr + 6);
-            }
-            set
-            {
-                setAt8(_rawvisiondata, ptr + 6, value);
-            }
-        }
-
-        public short dseg_2B65
-        {
-            get
-            {
-                return (short)getAt16(_rawvisiondata, ptr + 7);
-            }
-            set
-            {
-                setAt16(_rawvisiondata, ptr + 7, value);
-            }
-        }
-        public byte CameraY_2b66//;//2B66
-        {
-            get
-            {
-                return (byte)getAt8(_rawvisiondata, ptr + 8);
-            }
-            set
-            {
-                setAt8(_rawvisiondata, ptr + 8, value);
-            }
-        }
-
-        public TileInfo playerTileCopy_2B67; //2B67
-
-        public short dseg_2B6B
-        {
-            get
-            {
-                return (short)getAt16(_rawvisiondata, ptr + 0XD);
-            }
-            set
-            {
-                setAt16(_rawvisiondata, ptr + 0xD, value);
-            }
-        }
-        public short dseg_2B6D
-        {
-            get
-            {
-                return (short)getAt16(_rawvisiondata, ptr + 0XF);
-            }
-            set
-            {
-                setAt16(_rawvisiondata, ptr + 0xF, value);
-            }
-        }
-        // public static short dseg_2B6F;
-        // public static short FovYawXRIGHT = 0;//2b70        
-        // public static short FovYawYRIGHT = 0;//2b72
-
-        // public static short dseg_2B74;
-        // public static byte CameraX_2B75;
-        // public static byte dseg_2B76;
-        // public static byte CameraY_2B77;
-
-        // public static TileInfo playerTileCopy_2B78;
-        // public static byte dseg_2B7C;
-
-        int ptr;
-        static VisionParams()
-        {
-            for (int i = 0; i <= visionparams.GetUpperBound(0); i++)
-            {
-                visionparams[i] = new VisionParams(i);
-            }
-        }
-
-        VisionParams(int index)
-        {
-            ptr = index * 0x11;
-        }
-
-
-        /// <summary>
-        /// Port of a vanilla function which is used to calcuate what objects are inview. Implemented here to help support vanilla implementation of automapping
-        /// </summary>
-        public static void SetRangeOfVisionParams(short camerax, short cameray, short camerayaw)
-        {
-            var tileX = camerax >> 8; var tileY = cameray >> 8;
-            if (!UWTileMap.ValidTile(tileX, tileY))
-            {
-                return;
-            }
-            var tile = UWTileMap.current_tilemap.Tiles[tileX, tileY];
-            if (tile.tileType == 0)
-            {
-                RelatedToFov_2C60 = 0xF;
-            }
-            else
-            {
-                RelatedToFov_2C60 = 0;
-
-                visionparams[0].dseg_2B5E = 0x81;
-                visionparams[0].dseg_2B63 = 0;
-                visionparams[0].dseg_2B65 = 0;
-
-                visionparams[0].CameraX_2b64 = (byte)camerax;
-                visionparams[0].CameraY_2b66 = (byte)cameray;
-
-                visionparams[0].playerTileCopy_2B67 = UWTileMap.current_tilemap.Tiles[tileX, tileY];
-                //dseg_2B6B = reference to an array;
-                visionparams[1].dseg_2B5E = 0;//.dseg_2B6F = 0;
-                visionparams[1].dseg_2B63 = 0;//dseg_2B74 = 0;
-                visionparams[1].dseg_2B65 = 0;  //dseg_2B76 = 0;
-
-
-                visionparams[1].CameraX_2b64 = (byte)camerax; //CameraX_2B75 = (byte)camerax;
-                visionparams[1].CameraY_2b66 = (byte)cameray;//CameraY_2B77 = (byte)cameray;
-                visionparams[1].playerTileCopy_2B67 = UWTileMap.current_tilemap.Tiles[tileX, tileY];////playerTileCopy_2B78 = UWTileMap.current_tilemap.Tiles[tileX, tileY];
-                //dseg_2B7C =reference to an array that is set up in Seg032_6CF
-                short x= 0; short y = 0;
-                motion.SomethingProjectileHeading_seg021_22FD_EAE(heading: (ushort)(camerayaw + 0x2040), Result_arg2: ref x, Result_arg4: ref y);
-                visionparams[1].FovYawX = x;  visionparams[1].FovYawY = y;
-                motion.SomethingProjectileHeading_seg021_22FD_EAE(heading: (ushort)(camerayaw - 0x2040), Result_arg2: ref x, Result_arg4: ref y);
-                visionparams[0].FovYawX = x;  visionparams[0].FovYawY = y;
-                
-                visionparams[1].FovYawX = (short)(visionparams[1].FovYawX>>4);  
-                visionparams[1].FovYawY = (short)(visionparams[1].FovYawY>>4);
-                visionparams[0].FovYawX = (short)(visionparams[0].FovYawX>>4);  
-                visionparams[0].FovYawY = (short)(visionparams[0].FovYawY>>4);
-            }
-        }
-
-        public static void LikelyGetViewDistance()
-        {
-            var di = 0;
-            LikelyDistanceToWallOrDarkness = -1;
-            var currentshade = shade.shadesdata[playerdat.lightlevel].ShadingArray_26EF;
-            var var2currentshadePtr = 0;
-
-        seg032_1175:
-            LikelyDistanceToWallOrDarkness++;
-            //var var4 = RelatedToFov_2C60; //var4 appears to be a pointer to this value
-        seg032_119D:
-
-            if ((RelatedToFov_2C60 & 0xF) == 0xF)
-            {
-                //seg032_11B0
-                di = currentshade[66 + var2currentshadePtr];
-
-                //loop
-                seg032_11ED:
-                if ((RelatedToFov_2C60 & 0xF) == 0xF)
-                {
-                seg032_120B:
-                    if (currentshade[var2currentshadePtr] < di)
-                    {
-                        //seg032_11FD
-                        currentshade[var2currentshadePtr] = 0;
-                        var2currentshadePtr += 2;
-
-                        //Loop back to 120B
-                        goto seg032_120B;
-                    }
-                    else
-                    {
-                        //seg032_1210
-                        if (RelatedToFov_2C60 == 0xF)
-                        {
-                            return;
-                        }
-                        else
-                        {
-                            goto seg032_1175;
-                        }
-                    }
-                }
-                else
-                {
-                    seg032_11C8:
-                    if (visionparams[RelatedToFov_2C60].dseg_2B6D > currentshade[var2currentshadePtr])
-                    {
-                        //seg32_11BA
-                        currentshade[var2currentshadePtr] = 0;
-                        var2currentshadePtr +=2;
-                        goto seg032_11C8;
-                    }
-                    else
-                    {
-                        //seg32_11DE
-                        seg32_1014(ref RelatedToFov_2C60, currentshade, var2currentshadePtr);
-                        goto seg032_11ED;
-                    }
-                }
-            }
-            else
-            {
-                //seg032_1180
-                //lookup a dseg and call 
-                seg032_C9D(visionparams[0]);
-
-                goto seg032_119D;
-            }
-
-        }
-
-        static bool seg032_C9D(VisionParams vision)
-        {
-            return false;
-        }
-
-        static void seg32_1014(ref short ptrdseg_2c60, byte[] currentshade, int currentshade_ptr)
-        {
-            
-        }
-
-    }//end class
 }//end namespace
